disable cross origin

+++ src/main/java/ru/shop/demo/repository/PhotoRepository.java
@@ -4,7 +4,7 @@ import ru.shop.demo.models.Photo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface PhotoRepository extends JpaRepository<Photo, Integer>{
+public interface PhotoRepository extends JpaRepository<Photo, Long>{
 
     Optional<Photo> findById(long id);
